requirements.txt
asgiref==3.7.2
Django==4.2.5
django-autoslug==1.9.9
django-ckeditor==6.7.0
django_jazzmin==2.6.0
django-nested-inline==0.4.6
flake8==6.1.0
isort==5.12.0
pillow==10.0.1
psycopg2-binary==2.9.5
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.8.0